pyproject.toml
[tool.poetry]
name = "crewai-video-study-guide"
version = "1.0.0"
description = "Fast, AI-powered study guide generation from YouTube videos using CrewAI and Google Gemini"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
crewai = {extras = ["tools"], version = "^1.1.0"}
google-generativeai = "^0.8.5"
yt-dlp = "^2025.10.22"
opencv-python-headless = "^4.12.0"
youtube-transcript-api = "^1.2.3"
python-dotenv = "^1.1.1"
requests = "^2.32.5"
beautifulsoup4 = "^4.13.4"
lxml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^22.0.0"
flake8 = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
crewai-video-study = "src.main_crew:main"